attempt at reducing lines of code and cyclomatic complexity
